+++ js/grapheditor/EditorUi.js
@@ -2946,7 +2946,25 @@ EditorUi.prototype.initCanvas = function()
 				
 				mxEvent.consume(evt);
 			}), Editor.zoomFitImage, mxResources.get('fit'));
-	
+
+			/* 👇 SIYUAN 👇 */
+			// 添加取消灯箱模式按钮
+			if (window.mxIsSiyuan) {
+				addButton(
+					mxUtils.bind(
+						this,
+						function (evt) {
+							const url = new URL(window.location);
+							url.searchParams.delete('lightbox');
+							window.location.href = url.href;
+						}
+					),
+					Editor.thinDesignImage,
+					mxResources.get('edit'),
+				);
+			}
+			/* 👆 SIYUAN 👆 */
+
 			// Changes toolbar opacity on hover
 			var fadeThread = null;
 			var fadeThread2 = null;